+++ lib/services/src/orders_service.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:motocode_core/models/src/order_module.dart';
+import 'package:motocode_core/models/src/paginated_response.dart';
 import 'package:motocode_core/models/src/qr_code_module.dart';
 import 'package:motocode_core/services/src/api_client.dart';
 
@@ -34,6 +35,23 @@ class OrdersService {
         unawaited(_crashlytics?.recordError(error, StackTrace.current));
         throw error;
       });
+
+  Future<PaginatedResponse<Order>> getPaginatedOrders({
+    required int pageNumber,
+    required int pageSize,
+    String? searchQuery,
+  }) async {
+    return await _apiClient
+        .getPaginatedOrders(
+      pageNumber,
+      pageSize,
+      searchQuery,
+    )
+        .catchError((error) {
+      _crashlytics?.recordError(error, StackTrace.current);
+      throw error;
+    });
+  }
 
   /// Get order by purchase number
   ///